update annotations and readme
